app/src/androidTest/java/ca/ualberta/cs/lonelytwitter/TweetListTest.java
package ca.ualberta.cs.lonelytwitter;

import android.test.ActivityInstrumentationTestCase2;

import static android.R.id.list;
import static java.lang.Boolean.TRUE;

import java.util.ArrayList;
import java.util.Collections;
/**
 * Created by kivrikog on 10/12/17.
 */

public class TweetListTest extends ActivityInstrumentationTestCase2 {
    public TweetListTest(){
        super(ca.ualberta.cs.lonelytwitter.LonelyTwitterActivity.class);
    }
    public void testAddTweet(){
       // assertTrue(Boolean.TRUE);
        TweetList tweets = new TweetList();
        Tweet tweet = new NormalTweet("adding Tweet");
        tweets.add(tweet);
        assertTrue(tweets.hasTweet(tweet));

    }

    public void testDelete(){
        TweetList tweets=new TweetList();
        Tweet tweet = new NormalTweet("test");
        tweets.add(tweet);
        tweets.delete(tweet);
        assertFalse(tweets.hasTweet(tweet));
    }

    public void testHasTweet(){
        TweetList list= new TweetList();
        Tweet tweet = new NormalTweet("test");
        list.add(tweet);
        assertTrue(list.hasTweet(tweet));
    }

    public void testGetTweet(){
        TweetList list=new TweetList();
        Tweet tweet=new NormalTweet("test");
        list.add(tweet);
        //assertEquals(Boolean.TRUE);

    }
    public void testGetCount(){
        TweetList list = new TweetList();
        Tweet tweet = new NormalTweet("test");
        list.add(tweet);
        int getCount=list.getCount();
        assertEquals(getCount,1);
    }

}
